+++ crates/rl-docs/src/entries/stdlib/test/mod.rs
@@ -5,6 +5,7 @@ mod test_skip_if;
 mod test_assert_eq;
 mod test_assert_ne;
 mod test_assert_panics;
+mod test_run_registered;
 mod test_assert_no_panic;
 
 pub static TEST: StdEntry = StdEntry {
@@ -21,5 +22,6 @@ static FUNCTIONS: &[&FnEntry] = &[
     &test_assert_eq::TEST_ASSERT_EQ,
     &test_assert_ne::TEST_ASSERT_NE,
     &test_assert_panics::TEST_ASSERT_PANICS,
+    &test_run_registered::TEST_RUN_REGISTERED,
     &test_assert_no_panic::TEST_ASSERT_NO_PANIC,
 ];